requirements-test.txt
pytest>=7.0
GitPython==3.1.40
